# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Wedding Invitation API",
        "version": "1.0"
    },
    "paths": {
...
        "/api/templates": {
            "get": {
                "tags": [
                    "Templates"
                ],
                "summary": "Get paginated templates",
                "operationId": "515609062277ac1c04c633b7e0bad17e",
                "responses": {
                    "200": {
                        "description": "List of templates"
                    }
                }
            },
            "post": {
                "tags": [
                    "Templates"
                ],
                "summary": "Upload a new template",
                "operationId": "2d94858dd131018a268de19ec450647a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "title",
                                    "description",
                                    "type",
                                    "file"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Wedding Invitation"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Elegant wedding invitation template"
                                    },
                                    "type": {
                                        "type": "string",
                                        "enum": [
                                            "free",
                                            "premium"
                                        ],
                                        "example": "premium"
                                    },
                                    "file": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "minimum": 0,
                                        "example": 10.99
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Template uploaded successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/save-template": {
            "post": {
                "tags": [
                    "Templates"
                ],
                "summary": "Save a template for the authenticated user",
                "operationId": "1584f7bbf5ae3264227584ec1533726b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id"
                                ],
                                "properties": {
                                    "id": {
                                        "description": "The ID of the template to save",
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Template saved successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "409": {
                        "description": "Template already saved"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/templates/{id}": {
            "get": {
                "tags": [
                    "Templates"
                ],
                "summary": "Get a template by ID",
                "operationId": "c3970808bc05c3f90fab5309cbfd7ab4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Template details"
                    },
                    "404": {
                        "description": "Template not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Templates"
                ],
                "summary": "Update an existing template",
                "operationId": "f91a7687221b2bd95a1bceae8a778e43",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "description",
                                    "type",
                                    "flag_active"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Updated Template"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Updated description"
                                    },
                                    "type": {
                                        "type": "string",
                                        "enum": [
                                            "free",
                                            "premium"
                                        ],
                                        "example": "free"
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "minimum": 0,
                                        "example": 0
                                    },
                                    "flag_active": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Template updated successfully"
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Template not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/my-templates": {
            "get": {
                "tags": [
                    "Templates"
                ],
                "summary": "Get templates owned by the authenticated user",
                "operationId": "4f5ad80af4bb213c22c476d809780b63",
                "responses": {
                    "200": {
                        "description": "List of user's templates"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}